refactor: Extract copy to clipboard component
